+++ binstar_client/utils/__init__.py
@@ -15,7 +15,7 @@ from .spec import PackageSpec, package_specs, parse_specs
 
 # Re-export config
 from .config import (get_server_api, dirs, load_token, store_token, remove_token, get_config, set_config, load_config,
-                     USER_CONFIG, USER_LOGDIR, SITE_CONFIG, DEFAULT_CONFIG)
+                     USER_CONFIG, USER_LOGDIR, SITE_CONFIG, DEFAULT_CONFIG, get_binstar)
 
 from six.moves import input