+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## Fixed
 
+* Added back deprecated get_binstar function
 * Upload now reads Jupyter notebooks metadata
 
 ## Version 1.6.11 (2018/02/20)